Add ViewScheme separate file
